+++ settings.py
@@ -16,5 +16,6 @@ def positive_float_from_env(name, default):
 slack_token = os.environ.get('SLACK_TOKEN', '')
 messenger_token = os.environ.get('MESSENGER_TOKEN', '')
 messenger_verify_token = os.environ.get('MESSENGER_VERIFY_TOKEN', messenger_token)
+messenger_app_secret = os.environ.get('MESSENGER_APP_SECRET', '')
 messenger_url = "https://graph.facebook.com/v2.6/me/messages"
 request_timeout = positive_float_from_env('REQUEST_TIMEOUT', 5.0)